# Participium/api/openapi.yaml
openapi: 3.0.4
info:
  title: Participium - OpenAPI 1.0
  description: Participium Server based.
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
paths:
  /auth/users:
    post:
      tags:
        - Auth
      summary: Logs user into the system.
      description: Log into the system.
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: username
                password:
                  type: string
                  format: password
                  example: secure_word
              required:
                - username
                - password
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "400":
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /maintainers:
      post:
        tags:
          - Admin
        summary: Admin creates a maintainer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  password:
                    type: string
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/OfficeType"
                  active:
                    type: boolean
                required:
                  - name
                  - email
                  - password
                  - categories
        responses:
          "200":
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Maintainer"
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
      get:
        tags:
          - Maintainers
        summary: Officers lists all maintainers
        responses:
          "200":
            description: successful operation
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: "#/components/schemas/Maintainer"
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
...
  /auth/officers:
    post:
      tags:
        - Auth
      summary: Logs officers into the system.
      description: Log into the system.
      operationId: loginOfficer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: username
                password:
                  type: string
                  format: password
                  example: secure_word
              required:
                - username
                - password
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
                example: token
            application/xml:
              schema:
                type: string
                example: token
        "400":
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/maintainers:
    post:
      tags:
        - Auth
      summary: Logs maintainer into the system.
      description: Log into the system.
      operationId: loginMaintainer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: maintainer@example.com
                password:
                  type: string
                  format: password
                  example: secure_word
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string    
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "401":
          description: Invalid email/password supplied
        default:
          description: Unexpected error
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users:
    post:
      tags:
        - Users
      summary: Create user.
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users/logout:
    get:
      tags:
        - Users
      summary: Logs out current logged in user session.
      description: Log user out of the system.
      operationId: logoutUser
      parameters: []
      responses:
        "200":
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /officers:
    post:
      tags:
        - Admin
      summary: Admins create new Officer
      operationId: createOfficer
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Officer"
          application/xml:
            schema:
              $ref: "#/components/schemas/Officer"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Officer"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Officer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Officer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - Admin
      summary: Admin lists all officers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Officer"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/telegram:
    post:
      tags:
        - Auth
      summary: Retrieve user account associated with Telegram username
      description: Given a Telegram username, retrieve the corresponding user account from the system.
      operationId: retrieveTelegramAccount
      requestBody:
        description: Telegram username
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: telegram_user
              required:
                - username
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: user_token
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notifications:
    get:
      tags:
        - Notifications
      summary: List user notifications
      description: Returns notifications for the authenticated user. Use query param unreadOnly=true to filter unread.
      parameters:
        - in: query
          name: unreadOnly
          schema:
            type: boolean
          required: false
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /notifications/{id}/read:
    patch:
      tags:
        - Notifications
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  read:
                    type: boolean
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    OfficeType:
      type: string
      enum: [ "infrastructure", "environment","safety", "sanitation","transport","organization","other"]
    OfficerRole:
      type: string
      enum: ["municipal_public_relations_officer", "municipal_administrator","technical_office_staff"]
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: username
        firstName:
          type: string
          example: Mario
        lastName:
          type: string
          example: Rossi
        password:
          type: string
          example: placeholder
        email:
          type: string
          example: mail@domain.com
        avatar:
          type: string
          example: /uploads/avatars/167123123-avatar.jpg
        telegramUsername:
          type: string
          example: my_telegram
        emailNotifications:
          type: boolean
          example: true
        followedReports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
      xml:
        name: user
    Officer:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        roles:  # Array invece di singolo
          type: array
          items:
            type: object
            properties:
              officerRole:
                $ref: '#/components/schemas/OfficerRole'
              officeType:
                $ref: '#/components/schemas/OfficeType'
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        Coordinates:
          type: object
          properties:
            longitude:
              type: number
            latitude:
              type: number
    Report:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        Author:
          $ref: "#/components/schemas/User"
        Anonymity:
          type: boolean
        Date:
          type: string
        Category:
          $ref: "#/components/schemas/OfficeType"
        Document:
          type: object
          properties:
            Description:
              type: string
              minLength: 150
              example: the water is clear
            Photos:
              type: array
              items:
                type: string
              minItems: 1
              maxItems: 3
        assignedMaintainerId:
          type: integer
          nullable: true
        followerUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        reportId:
          type: integer
        type:
          type: string
          enum:
            - STATUS_CHANGE
            - OFFICER_MESSAGE
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        read:
          type: boolean
    Maintainer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        password:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/OfficeType"
        active:
          type: boolean
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header