# Participium/server/src/models/dto/Officer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Participium - OpenAPI 1.0
 * Participium Server based.
 *
 * The version of the OpenAPI document: 1.0.12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../../../generated/runtime';
import {
    OfficerRoleFromJSON,
    OfficerRoleToJSON,
} from './OfficerRole';
import type { OfficerRole } from './OfficerRole';
import type { OfficeType } from './OfficeType';
import {
    OfficeTypeFromJSON,
    OfficeTypeToJSON,
} from './OfficeType';


/**
 * Entry per ruolo e ufficio dell'Officer
 * @export
 * @interface RoleEntry
 */
export interface RoleEntry {
    role: OfficerRole;
    office: OfficeType | null; // consenti null per compatibilità con repository
}

/**
 * 
 * @export
 * @interface Officer
 */
export interface Officer {
    id?: number;
    username?: string;
    name?: string;
    surname?: string;
    email?: string;
    password?: string;
    /**
     * Lista di ruoli associati all'Officer
     */
    roles?: RoleEntry[];
}



/**
 * Check if a given object implements the Officer interface.
 */
export function instanceOfOfficer(value: object): value is Officer {
    return true;
}

export function OfficerFromJSON(json: any): Officer {
    return OfficerFromJSONTyped(json, false);
}

export function OfficerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Officer {
    if (json == null) {
        return json;
    }

    // Supporto input legacy con singolo Role/Office
    const legacyRole = json['Role'] == null ? undefined : OfficerRoleFromJSON(json['Role']);
    const legacyOffice = json['Office'] == null ? undefined : OfficeTypeFromJSON(json['Office']);

    // Supporto input moderno con array "roles"
    const roles = Array.isArray(json['roles'])
        ? json['roles'].map((r: any) => ({
            role: OfficerRoleFromJSON(r.role),
            office: r.office == null ? null : OfficeTypeFromJSON(r.office),
        }))
        : [];

    return {
        'id': json['id'] == null ? undefined : json['id'],
        'username': json['username'] == null ? undefined : json['username'],
        'name': json['name'] == null ? undefined : json['name'],
        'surname': json['surname'] == null ? undefined : json['surname'],
        'email': json['email'] == null ? undefined : json['email'],
        'password': json['password'] == null ? undefined : json['password'],
        'roles': roles ?? (legacyRole != null && legacyOffice != null ? [{ role: legacyRole, office: legacyOffice }] : undefined),
    };
}

export function OfficerToJSON(json: any): Officer {
    return OfficerToJSONTyped(json, false);
}

export function OfficerToJSONTyped(value?: Officer | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) return value;
    return {
        'id': value['id'],
        'username': value['username'],
        'name': value['name'],
        'surname': value['surname'],
        'email': value['email'],
        'password': value['password'],
        // output moderno: array roles
        roles: value.roles?.map(r => ({
          role: OfficerRoleToJSON(r.role),
          office: r.office == null ? null : OfficeTypeToJSON(r.office),
        })),
        // opzionale: se serve compatibilità con API legacy, puoi anche mantenere:
        // 'Role': value['roles']?.[0] ? OfficerRoleToJSON(value['roles'][0].role) : undefined,
        // 'Office': value['roles']?.[0] ? OfficeTypeToJSON(value['roles'][0].office) : undefined,
    };
}